Restore float return in PhaseDelay case
Use DOUBLES_EQUAL rather than LONGS_EQUAL

+++ PhaseDelay-include/PhaseDelay.h
@@ -1,7 +1,7 @@
 #ifndef EXAMPLE_INCLUDED
 #define EXAMPLE_INCLUDED
 
-int getPhaseDelayFromRawSamples(void);
+float getPhaseDelayFromRawSamples(void);
 
 #endif
