+++ PhaseDelay-src/PhaseDelay.c
@@ -11,8 +11,8 @@ uint16_t raw_samples[256]; // The values are from 0 to 4096
 * When the function is called, we assume that raw_samples
 * already has the most up-to-date samples from the ADC.
 */
-int getPhaseDelayFromRawSamples(void) {
-  int theta = 1; // TODO: Restore to returning a float
+float getPhaseDelayFromRawSamples(void) {
+  float theta = 1.0;
 
  // TODO: Please implement this function
  return theta; // theta is between 0 and 360
